+++ src/views/trees/numberTreeItem.ts
@@ -10,16 +10,21 @@ export class NumberTreeItem extends BaseTreeItem {
 
   constructor(
     public number: INumber) {
-    super(number.msisdn, TreeItemCollapsibleState.None, 'symbol-number', undefined, 'number');
+    super(`+${number.msisdn}`, TreeItemCollapsibleState.None, 'active-call-line', undefined, 'number');
     this.setDescription();
     this.setToolTip();
   }
+
+  public setContextValue(): void {
+    this.contextValue = `number${Object.keys(this.number.features).length > 0 ? '-' + Object.keys(this.number.features).join('-') : ''}`;
+  }
+
 
   private setDescription(): void {
     this.description = this.number.features.join(' ').toLowerCase();
   }
 
-  private setToolTip() {
+  private setToolTip(): void {
     this.tooltip = resolveTooltipMarkdown(this.numberTooltipTemplate, { ...this.number });
   }